+++ src/main/java/com/danakga/webservice/board/service/BoardService.java
@@ -1,12 +1,9 @@
 package com.danakga.webservice.board.service;
 
 import com.danakga.webservice.board.dto.request.ReqBoardWriteDto;
-import com.danakga.webservice.board.dto.request.ReqFileUploadDto;
-import com.danakga.webservice.board.dto.response.ResBoardUpdateDto;
-import com.danakga.webservice.board.dto.response.ResBoardWriteDto;
+import com.danakga.webservice.board.dto.response.ResBoardDto;
 import com.danakga.webservice.board.model.Board;
 import com.danakga.webservice.user.model.UserInfo;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
@@ -16,9 +13,9 @@ public interface BoardService {
     List<Board> list();
 
     //게시글 작성
-    ResBoardWriteDto write(ReqBoardWriteDto reqBoardWriteDto, ReqFileUploadDto reqFileUploadDto, MultipartFile files, UserInfo userInfo);
+    ResBoardDto write(ReqBoardWriteDto reqBoardWriteDto, UserInfo userInfo);
 
 
-//    //게시글 수정
-//    ResBoardUpdateDto update(Board board);
+    //게시글 수정
+    ResBoardDto.ResBoardUpdateDto update(Board board);
 }
