+++ src/main/java/com/danakga/webservice/board/service/FileService.java
@@ -1,11 +1,12 @@
 package com.danakga.webservice.board.service;
 
 import com.danakga.webservice.board.dto.request.ReqFileUploadDto;
-import com.danakga.webservice.board.dto.response.ResFileUploadDto;
+import com.danakga.webservice.board.dto.response.ResBoardDto;
 import com.danakga.webservice.board.model.Board;
-import com.danakga.webservice.board.repository.FileRepository;
+import org.springframework.web.multipart.MultipartFile;
 
 public interface FileService {
 
-    ResFileUploadDto savefile(ReqFileUploadDto reqFileUploadDto, Board board);
+    //파일 업로드 후 저장
+    ResBoardDto.ResFileUploadDto saveFile(ReqFileUploadDto reqFileUploadDto, MultipartFile files, Board board);
 }
